requirements.txt
discord.py
configparser
argparse
